backend/requirement.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv
python-multipart
gTTS
email-validator

